feat: most recent selector
