+++ catalyst_voices/packages/internal/catalyst_voices_blocs/lib/src/discovery/discovery_state.dart
@@ -3,19 +3,23 @@ part of 'discovery_cubit.dart';
 final class DiscoveryState extends Equatable {
   final DiscoveryCurrentCampaignState currentCampaign;
   final DiscoveryCampaignCategoriesState campaignCategories;
+  final DiscoveryMostRecentProposalsState mostRecentProposals;
 
   const DiscoveryState({
     this.currentCampaign = const DiscoveryCurrentCampaignState(),
     this.campaignCategories = const DiscoveryCampaignCategoriesState(),
+    this.mostRecentProposals = const DiscoveryMostRecentProposalsState(),
   });
 
   DiscoveryState copyWith({
     DiscoveryCurrentCampaignState? currentCampaign,
     DiscoveryCampaignCategoriesState? campaignCategories,
+    DiscoveryMostRecentProposalsState? mostRecentProposals,
   }) {
     return DiscoveryState(
       currentCampaign: currentCampaign ?? this.currentCampaign,
       campaignCategories: campaignCategories ?? this.campaignCategories,
+      mostRecentProposals: mostRecentProposals ?? this.mostRecentProposals,
     );
   }
 
@@ -23,6 +27,7 @@ final class DiscoveryState extends Equatable {
   List<Object?> get props => [
         currentCampaign,
         campaignCategories,
+        mostRecentProposals,
       ];
 }
 
@@ -72,5 +77,30 @@ final class DiscoveryCampaignCategoriesState extends Equatable {
         isLoading,
         error,
         categories,
+      ];
+}
+
+final class DiscoveryMostRecentProposalsState extends Equatable {
+  final bool isLoading;
+  final LocalizedException? error;
+  final List<PendingProposal> proposals;
+
+  const DiscoveryMostRecentProposalsState({
+    this.isLoading = true,
+    this.error,
+    this.proposals = const [],
+  });
+
+  bool get showProposals => !isLoading && proposals.isNotEmpty && error == null;
+
+  bool get showEmptyState => !isLoading && proposals.isEmpty && error == null;
+
+  bool get showError => !isLoading && error != null;
+
+  @override
+  List<Object?> get props => [
+        isLoading,
+        error,
+        proposals,
       ];
 }
